UI for building and location selection
all thats left in terms of ui is that location selection needs to go to form that will use URL parameters to make sure it sends report to the proper coresponsing location clicked on. All cards and location are dynamically rendered as data is added to backend they should appear in front end. For the time being I made dummy data to display the dynamic rendering until we have backend fully ready. Then that data will instead be fetched from API