- report connected
Report connected to each spot individually so when making report user will not have to input spot they want it to be for, it will just automatically submit data to that spot based on the one clicked on. That is not fully realized but i have a plan for that which is... each form is dynamically linked to each spot meaning each form has its own URL based on the building and spot picked for it. useParams hook will be used to grab the buildingID and the spotID from the link and thats how that specific report will know where to be sent to. Also the report has been styled.,